fixed bugs related to edit delete and creation on trips

+++ app/controllers/TripController.php
@@ -18,6 +18,28 @@ class TripController extends Controller
     public function create()
     {
         $formData = $this->request->getData();
+
+        if(BusRepository::readById($formData['busId']) === null)
+        {
+            $res = new Response('There is no bus with that ID!',403);
+            $res->send();
+            exit();
+        }
+
+        if(LocationRepository::readById($formData['departureId']) === null)
+        {
+            $res = new Response('There is no location with that ID!',403);
+            $res->send();
+            exit();
+        }
+
+        if(LocationRepository::readById($formData['arrivalId']) === null)
+        {
+            $res = new Response('There is no location with that ID!',403);
+            $res->send();
+            exit();
+        }
+
         $busId = $formData['busId'];
         $departureId = $formData['departureId'];
         $arrivalId = $formData['arrivalId'];
@@ -43,9 +65,24 @@ class TripController extends Controller
     {
         $formData = $this->request->getData();
         $id = $formData['id'];
+        $trip = TripRepository::readById($id);
 
-        $trip = $this->viewData['tripRepo'][$id];
-        
+        $checkBookings = BookingRepository::countByTrip($trip->getId());
+
+        if($checkBookings > 0)
+        {
+            $res = new Response("For this trip exists bookings. You can't delete it before deleting the bookings associated with it!",500);
+            $res->send();
+            exit();
+        }
+
+        if($trip === null)
+        {
+            $res = new Response('There is no location with that ID!',403);
+            $res->send();
+            exit();
+        }
+
         $checkDelete = TripRepository::delete($trip);
 
         if($checkDelete === false)
@@ -64,7 +101,7 @@ class TripController extends Controller
     {
         $formData = $this->request->getData();
         $id = $formData['id'];
-        $trip = $this->viewData['tripRepo'][$id];
+        $trip = TripRepository::readById($id);
 
         if(isset($trip) && !empty($trip))
         {
@@ -83,13 +120,43 @@ class TripController extends Controller
         $formData = $this->request->getData();
 
         $id = $formData['id'];
+
+        if(BusRepository::readById($formData['busId']) === null)
+        {
+            $res = new Response('There is no bus with that ID!',403);
+            $res->send();
+            exit();
+        }
+
+        if(LocationRepository::readById($formData['departureId']) === null)
+        {
+            $res = new Response('There is no location with that ID!',403);
+            $res->send();
+            exit();
+        }
+
+        if(LocationRepository::readById($formData['arrivalId']) === null)
+        {
+            $res = new Response('There is no location with that ID!',403);
+            $res->send();
+            exit();
+        }
+
         $busId = $formData['busId'];
         $departureId = $formData['departureId'];
         $arrivalId = $formData['arrivalId'];
         $dateTimeStart = $formData['dateTimeStart'];
         $dateTimeEnd = $formData['dateTimeEnd'];
 
-        $trip = $this->viewData['tripRepo'][$id];
+        $trip = TripRepository::readById($id);
+
+        if($trip === null)
+        {
+            $res = new Response('There is no trip with that ID!',403);
+            $res->send();
+            exit();
+        }
+
         $trip->setBusId($busId);
         $trip->setLocationStartId($departureId);
         $trip->setLocationEndId($arrivalId);
